build/windows: Add optional one-arch support to Windows installer
This removes the old custom DIR DDIR code from the era the
installer was built locally. Now, it refers to bundles and
will make the .exe installer based on the bundles avaiable,
not enforcing the developer to download all bundles anymore.

We will not use such new facility on CI since we distribute an
universal installer with all arches (for very good reasons).

The main advantage is locally. Now the .iss script can
recognize at ISCC build time the bundles on the machine,
to create, for example, an one-arch installer. This is very
fast and practival for local hacking the installer scripts.

Of course, this also makes the installer script powerful and
consistent with all other dist scripts which already can do
one-arch packages (appimage, flatpak, snap and msix).