images/logo: Fix redundant Gimp.Procedure.run calls
Mitigates GNOME/gimp#11385

In some scripts Gimp.Procedure.run was being run twice in the same proc, which
is wrong and could be contributing to the tendency of stuckness.

Also, we were sending EX_SOFTWARE exit code on failures, but it does
not exist on Windows, it is UNIX-only. This commit fixes that too.

(cherry picked from commit 0438699665d8fa711a8537c61497b0928446358f)