app: new infrastructure to pass messages to GIMP installations.
This simply reuses the gimp_versions.json file, which is loaded no more
than once a week on Windows and macOS. It doesn't do any additional HTTP
request.
Furthermore it stays a read-only request (not sending any information to
upstream which is anyway a static website).

I had the idea of this feature for quite some time actually, but lately
I came to realize it would have been very handy right now if it had been
implemented. Indeed we have this issue where GIMP is crashing on
Norwegian Bokmål or Turkish (#12626) and this has lasted for a few
months now.  Unfortunately, while some people would know where to find
us (on the bug tracker), most people would just be stuck and frustrated
with a broken GIMP without any information. It would have been great if
we had been able to at least display a message to Windows users and give
them some workaround advice, or at least some info for them to be a bit
patient.

We won't over-use this feature, but for the few off times where we could
need it, it would be extremely useful IMO.

What remains to be done:

* Split a bit the logic so that this can work also on Linux, or on the
  store version for Windows, i.e. when we don't actually check for new
  versions (yet we could still check for messages).
* Add a new settings so that people can disable this message check
  independently to the version check? Though it doesn't bring much to be
  fair, since if it does not do any more requests with 1 or 2 checks.
* I need to implement some condition logic for messages which need to be
  seen only on specific OSes or with specific versions, etc.
* Maybe implement some localization for messages?