gimp-ux#176: align expander arrow to the right.
The GimpCheckExpander now takes the max possible amount as allocated to
the full frame. It means that if several GimpCheckExpander objects are
added on a same container and are all made to take as much space as
possible, their arrow should also align.

As a secondary update in this commit, the full space after the label is
clickable and would expand the frame (not just the arrow), hence making
expanding/collapsing easier without having to be extra accurate in
pointing the small arrow.