app: Use 'g_message' (like the help plug-in) when GVFSd-HTTP is not present
Since GVFS is listed in the INSTALL file as a required dependency, it is
fair enough to be more clear when it is not present, otherwise users will
have outdated GIMP without knowing like happened in macOS in the past.

Also, the check update button needs some feedback when it fails by the way.

Closes #issue_number