desktop: add a <url> tag to our <release> tag.
I believe this is used in software listings (GNOME Software, KDE
Discovery, Flathub…) as a link to get more information. So it would be
quite useful to redirect people who want to know more towards our
website.

Up to now, we could not use this AppStream feature because we were never
completely sure of the final release news URL (because of the date part
in particular, since we were often publishing the news a day or 2 after
the tagging). With the new alias feature installed on gimp-web, we now
have an easy to remember and unchanging alias. Let's use this!