build/linux: Do not output version on Snap dist script (rev takes precedence)
Unlike the AppImage, .exe installer and .msix, the meson 'version' info is
not relevant on snap versioning. It is mostly managed by 'rev' info, which
is similar to 'commit' on Flatpak. Such rev information is not set by us, is
set externally by snapd for local .snap, and by Snap Store for remote .snap.