Revert "desktop: Make clear that GIMP is positioned as professional program"
This reverts commit c4b4f343d4284c801dbe8258ae6cdf28d2d1a603.

Though I did obviously say such things, I think we must think more
thoroughly on how we want to word it, since it is the entry point to our
software. So it's pretty important.

Also using the word "professional" in particular is problematic as we
seem to exclude various amateurs (in the laudative sense of the term) of
high skills. Wording such as "advanced", "expert" or "high end" may be
more appropriate.

Anyway reverting until we make a team decision for better wording.