plug-ins: determine correctly whether to export a raster object or not.
There were 3 bugs which were kinda countering each other and
participating to some confusion regarding actual layer types to export.

First the previous commit as we were looking whether a layer had been
rasterized to determine the object type on libgimp side. Secondly this
commit as the plug-in was using the previously wrong procedures. Lastly
with the upcoming commit, when we were duplicating an image, we were
disregarding any rasterization.