libgimpwidgets: consider NULL value of GimpLabelEntry as empty string.
Fixes:

> (file-svg:125356): Gtk-CRITICAL **: 22:11:43.459: gtk_entry_buffer_set_text: assertion 'chars != NULL' failed

… when running file-svg-export because the default of the "title" string
argument is NULL.

I did hesitate considering this was a bug in file-svg code on the
premises that a NULL value was not valid (hence replacing it as ""), but
in the end, I went with the more flexible solution, which is that NULL
is considered equivalent to "". So GimpLabelEntry code will just
transform the NULL pointer internally to an empty string.

(cherry picked from commit faa288300fcf1b58398399c635cfac52e5ea7fff)