Issue #11613: crashing when freeing window handle.
It is a tentative fix as I had this crash once but could not reproduce
it, even redoing dozens of times the same thing I had done.

It is mostly based on something which GDK docs of
gdk_wayland_window_export_handle() says:

> To unexport the window, gdk_wayland_window_unexport_handle() must be
> called the same number of times as gdk_wayland_window_export_handle()
> was called. Any 'exported' callback may still be invoked until the
> window is unexported or destroyed.

So when the GdkWindow of the widget still exists, let's make sure we
call gdk_wayland_window_unexport_handle() as many times as necessary (no
less, but also no more!), and also that we disconnect all handlers which
could possibly call more export_handle().