libgimpconfig: better error message when a GObject property tries to…
… deserialize with a wrong or unknown type.

The error message in issue #13787 was about trying to deserialize a
GimpControllerMouse object as GimpControllerInfo's controller. Yet
GimpControllerMouse was removed with commit 76ddf4421c so this was
failing.

With this change, such error would be more explicit, with an error
saying:

> Unknown object type: GimpControllerMouse

… instead of:

> unexpected character ')', expecting string constant

(which was very confusing to the actual issue)

(cherry picked from commit 5831b7ddcbbf78a7f6994f631eae0da3b3276ddc)